+++ ui/options_popup.gd
@@ -94,6 +94,9 @@ extends PopupPanel
 	],
 	LABEL_GRAPHICS_PERFORMANCE = [
 		[&"LABEL_STARMAP", &"starmap"],
+		[&"LABEL_MSAA", &"msaa_3d"],
+		[&"LABEL_FXAA", &"fxaa"],
+		[&"LABEL_TAA", &"use_taa"],
 	],
 }
 
@@ -106,6 +109,7 @@ extends PopupPanel
 	language = [&"LanguageManager", &"language_settings"],
 	gui_size = [&"CoreSettings", &"gui_size_settings"],
 	starmap = [&"Global", &"StarmapSize"],
+	msaa_3d = [&"GraphicsManager", &"msaa_settings"],
 }
 
 ## Each option Control can be a [CheckBox], [OptionButton], [SpinBox],
@@ -156,6 +160,15 @@ func _configure_after_core_inited() -> void:
 	if autoremove_for_missing_save_plugin and !IVPluginUtils.is_plugin_enabled("ivoyager_save"):
 		for column in layout:
 			column.erase(&"LABEL_SAVE_LOAD")
+	if IVGlobal.is_gl_compatibility:
+		# FXAA and TAA are unsupported in the Compatibility renderer (incl. web).
+		var graphics_section: Array = section_content[&"LABEL_GRAPHICS_PERFORMANCE"]
+		var supported_options: Array = []
+		for option_array: Array in graphics_section:
+			var setting: StringName = option_array[1]
+			if setting != &"fxaa" and setting != &"use_taa":
+				supported_options.append(option_array)
+		section_content[&"LABEL_GRAPHICS_PERFORMANCE"] = supported_options
 
 
 func _shortcut_input(event: InputEvent) -> void: